opensearch-index-build-logs/requirements.txt
boto3==1.24.85
botocore==1.27.85
kubernetes==25.3.0
pyjson5==1.6.2
python-gitlab==3.11.0
requests==2.28.1
